+++ pysui/abstracts/client_types.py
@@ -51,7 +51,9 @@ class AbstractType(ABC):
         if isinstance(self.value, bytes):
             return self.value.decode()
         if isinstance(self.value, int):
-            return self.value
+            return str(self.value)
+        if isinstance(self.value, bool):
+            return "true" if self.value else "false"
         if isinstance(self.value, AbstractType):
             return str(self.value.value)
         if isinstance(self.value, NoneType):