+++ pysui/sui/sui_bcs.py
@@ -14,15 +14,21 @@
 """Sui BCS Conversions [EXPERIMENTAL]."""
 
 import base64
-import binascii
-from typing import Any, Union
+from functools import partial
+import math
+from typing import Union
 import canoser
+
+# pylint:disable=wildcard-import,unused-wildcard-import
+from pysui.sui.sui_types.bcs import *
+
+# pylint:enable=wildcard-import,unused-wildcard-import
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_utils import hexstring_to_list, b58str_to_list, b64str_to_list
 from pysui.sui.sui_clients.common import SuiRpcResult
 from pysui.sui.sui_txresults.common import GenericRef
 from pysui.sui.sui_txresults.single_tx import ObjectRawRead
-from pysui.sui.sui_builders.get_builders import GetRawObject
+from pysui.sui.sui_builders.get_builders import GetFunction, GetRawObject
 from pysui.sui.sui_builders.exec_builders import (
     _MoveCallTransactionBuilder,
     MoveCall,
@@ -34,10 +40,9 @@ from pysui.sui.sui_builders.exec_builders import (
 from pysui.sui.sui_clients.sync_client import SuiClient as SyncClient
 from pysui.sui.sui_types.scalars import ObjectID, SuiString
 from pysui.sui.sui_types.collections import SuiArray
+from pysui.sui.sui_txresults.package_meta import SuiMoveFunction, SuiParameterReference, SuiParameterStruct
 
 
-_BCS_ADDRESS_LENGTH: int = 20
-_BCS_DIGEST_LENGTH: int = 32
 _TKIND_INDEX: int = 0
 _SUB_TKIND_INDEX: int = 1
 _SKIP_KIND_AND_SINGLE: int = 2
@@ -46,356 +51,15 @@ _GAS_AND_BUDGET_BYTE_OFFSET: int = -97
 FAKE_ADDRESS_OR_OBJECT: str = "0x0000000000000000000000000000000000000000"
 
 
-class BCSAddress(canoser.Struct):
-    """BCSAddress Represents a Sui Address or ObjectID in int list format.
-
-    :ivar _fields: Contains the Address BCS descriptor
-    :ivar type _fields: list[int]
-    """
-
-    _fields = [("Address", canoser.ArrayT(canoser.Uint8, _BCS_ADDRESS_LENGTH, False))]
-
-    def to_str(self) -> str:
-        """."""
-        return binascii.hexlify(bytes(getattr(self, "Address"))).decode()
-
-    def to_address_str(self) -> str:
-        """."""
-        return f"0x{self.to_str()}"
-
-    def to_sui_address(self) -> SuiAddress:
-        """."""
-        return SuiAddress(self.to_address_str())
-
-    @classmethod
-    def from_sui_address(cls, indata: SuiAddress) -> "BCSAddress":
-        """."""
-        return cls(hexstring_to_list(indata.address))
-
-    @classmethod
-    def from_str(cls, indata: str) -> "BCSAddress":
-        """."""
-        return cls(hexstring_to_list(indata))
-
-
-class BCSDigest(canoser.Struct):
-    """BCSDigest represents a transaction or object base58 value.
-
-    :ivar _fields: Contains the Digest BCS descriptor
-    :ivar type _fields: list[int]
-    """
-
-    _fields = [("Digest", canoser.ArrayT(canoser.Uint8, _BCS_DIGEST_LENGTH))]
-
-    @classmethod
-    def from_str(cls, indata: str) -> "BCSDigest":
-        """."""
-        return cls(b58str_to_list(indata))
-
-
-class BCSObjectReference(canoser.Struct):
-    """BCSObjectReference represents an objects reference data.
-
-    :ivar _fields: Contains the ObjectID,SequenceNumber and Digest BCS descriptor
-    :ivar type _fields: list[int]
-    """
-
-    _fields = [
-        ("ObjectID", BCSAddress),
-        ("SequenceNumber", canoser.Uint64),
-        ("ObjectDigest", BCSDigest),
-    ]
-
-    @classmethod
-    def from_generic_ref(cls, indata: GenericRef) -> "BCSObjectReference":
-        """."""
-        return cls(BCSAddress.from_str(indata.object_id), indata.version, BCSDigest.from_str(indata.digest))
-
-
-class BCSSharedObjectReference(canoser.Struct):
-    """."""
-
-    _fields = [
-        ("ObjectID", BCSAddress),  # canoser.ArrayT(canoser.Uint8, _BCS_ADDRESS_LENGTH)),
-        ("SequenceNumber", canoser.Uint64),
-        ("Mutable", bool),
-    ]
-
-    @classmethod
-    def from_generic_ref(cls, indata: GenericRef) -> "BCSSharedObjectReference":
-        """."""
-        return cls(BCSAddress.from_str(indata.object_id), indata.version, BCSDigest.from_str(indata.digest))
-
-
-class BCSOptionalU64(canoser.RustOptional):
-    """BCSOptionalU64 Optional assignment of unsigned 64 bit int.
-
-    :ivar _type: Identifies this object as unsigned 64 bit int.
-    :ivar type _type: canoser.Uint64
-    """
-
-    _type = canoser.Uint64
-
-
-class Uint256(canoser.int_type.IntType):
-    """Uint256 represents a 256 bit ulong as hack as canoser doesn't support."""
-
-    byte_lens = 32
-    max_value = 115792089237316195423570985008687907853269984665640564039457584007913129639935
-    min_value = 0
-    signed = False
-
-    @classmethod
-    def encode(cls, value):
-        """."""
-        return value.to_bytes(32, byteorder="little", signed=False)
-
-
-class TypeTag(canoser.RustEnum):
-    """TypeTag enum for move call type_arguments."""
-
-    _enums = [
-        ("Bool", bool),
-        ("U8", canoser.Uint8),
-        ("U64", canoser.Uint64),
-        ("U128", canoser.Uint128),
-        ("Address", BCSAddress),
-        ("Signer", None),
-        ("Vector", None),  # Injected below StructTag
-        ("Struct", None),  # Injected below StructTag
-        ("U16", canoser.Uint16),
-        ("U32", canoser.Uint32),
-        ("U256", Uint256),
-    ]
-
-    @classmethod
-    def update_value_at(cls, index: int, value: Any):
-        """."""
-        cls._enums[index] = (cls._enums[index][0], value)
-
-
-class StructTag(canoser.Struct):
-    """StructTag."""
-
-    _fields = [("address", BCSAddress), ("module", str), ("name", str), ("type_parameters", [TypeTag])]
-
-    @classmethod
-    def from_type_str(cls, type_str: str) -> "StructTag":
-        """from_type_str convert a type_arg to StructTag.
-
-        :param type_str: Type string (e.g. 0x2::sui::SUI)
-        :type type_str: str
-        :return: Instance of StructTag
-        :rtype: StructTag
-        """
-        if type_str.count("::") == 2:
-            split_type = type_str.split("::")
-            return cls(BCSAddress.from_str(split_type[0]), split_type[1], split_type[2], [])
-        raise ValueError(f"Ill formed type_argument {type_str}")
-
-
-# Overcome forward reference at init time with these injections
-TypeTag.update_value_at(6, [TypeTag])
-TypeTag.update_value_at(7, StructTag)
-
-
-class ObjectArg(canoser.RustEnum):
-    """ObjectArg."""
-
-    _enums = [("ImmOrOwnedObject", BCSObjectReference), ("SharedObject", BCSSharedObjectReference)]
-
-
-class CallArg(canoser.RustEnum):
-    """CallArg."""
-
-    _enums = [("Pure", [canoser.Uint8]), ("Object", ObjectArg), ("ObjVec", [ObjectArg])]
-
-
-class BCSMoveCall(canoser.Struct):
-    """BCSMoveCall represents a sui_moveCall structure."""
-
-    _fields = [
-        ("package", BCSAddress),
-        ("module", str),
-        ("function", str),
-        ("type_arguments", [TypeTag]),
-        ("arguments", [CallArg]),
-    ]
-
-
-class BCSPay(canoser.Struct):
-    """BCSPay represents a sui_pay structure."""
-
-    _fields = [
-        ("coins", [BCSObjectReference]),
-        ("recipients", [BCSAddress]),
-        ("amounts", [canoser.Uint64]),
-    ]
-
-
-class BCSPaySui(canoser.Struct):
-    """BCSPaySui represents a sui_paySui structure.
-
-    **Not Supported in sui_devInspectTransaction yet.**
-    """
-
-    _fields = [
-        ("coins", [BCSObjectReference]),
-        ("recipients", [BCSAddress]),
-        ("amounts", [canoser.Uint64]),
-    ]
-
-
-class BCSPayAllSui(canoser.Struct):
-    """BCSPayAllSui represents a sui_payAllSui structure.
-
-    **Not Supported in sui_devInspectTransaction yet.**
-    """
-
-    _fields = [
-        ("coins", [BCSObjectReference]),
-        ("recipient", BCSAddress),
-    ]
-
-
-class BCSTransferObject(canoser.Struct):
-    """BCSTransferObject represents a sui_transferObjecrt structure."""
-
-    _fields = [
-        ("recipient", BCSAddress),
-        ("object_ref", BCSObjectReference),
-    ]
-
-
-class BCSTransferSui(canoser.Struct):
-    """BCSTransferSui represents a sui_transferSui structure."""
-
-    _fields = [
-        ("recipient", BCSAddress),
-        ("amount", BCSOptionalU64),
-    ]
-
-
-class BCSPublish(canoser.Struct):
-    """BCSPublish represents a sui_publish structure.
-
-    **Not Supported in sui_devInspectTransaction yet.**
-    """
-
-    _fields = [("modules", canoser.ArrayT(canoser.ArrayT(canoser.Uint8)))]
-
-
-class BCSSingleTransaction(canoser.RustEnum):
-    """BCSSingleTransaction is enumeration of different single (not batch) transactions."""
-
-    _enums = [
-        ("TransferObject", BCSTransferObject),  # Works
-        ("Publish", BCSPublish),  # Not Works
-        ("Call", BCSMoveCall),  # Works
-        ("TransferSui", BCSTransferSui),  # Works
-        ("Pay", BCSPay),  # Works
-        ("PaySui", BCSPaySui),  # Not work
-        ("PayAllSui", BCSPayAllSui),  # Not work
-        ("ChangeEpoch", None),  # Not implemented
-        ("Genesis", None),  # Not implemented
-    ]
-
-    @classmethod
-    def variant_for_index(cls, index: int) -> Union[tuple[str, canoser.RustEnum], ValueError]:
-        """."""
-        if index > len(cls._enums):
-            raise ValueError(f"{cls.__name__} has only {len(cls._enums)} and index requested is greater {index}")
-        return cls._enums[index]
-
-
-class BCSBatchTransaction(canoser.Struct):
-    """BCSBatchTransaction represents a sui_batchTransaction structure."""
-
-    _fields = [("Singles", [BCSSingleTransaction])]
-
-
-class BCSTransactionKind(canoser.RustEnum):
-    """BCSTransactionKind is enumeration of different transaction types."""
-
-    _enums = [
-        ("Single", BCSSingleTransaction),
-        ("Batch", BCSBatchTransaction),  # Not implemented
-    ]
-
-    @classmethod
-    def variant_for_index(cls, index: int) -> Union[tuple[str, canoser.RustEnum], ValueError]:
-        """variant_for_index returns the name and BCS type at enum index.
-
-        :param index: The index of which TransactionKind enum varient to return
-        :type index: int
-        :raises ValueError: Out of range index
-        :return: The variant name and value at index
-        :rtype: Union[tuple[str, canoser.RustEnum], ValueError]
-        """
-        if index > len(cls._enums):
-            raise ValueError(f"{cls.__name__} has only {len(cls._enums)} and index requested is greater {index}")
-        return cls._enums[index]
-
-
-def tkind_from_result(indata: SuiRpcResult) -> Union[str, SuiRpcResult]:
-    """tkind_from_result Return a BCS serialized kind as base64 encoded string.
-
-    :param indata: _description_
-    :type indata: SuiRpcResult
-    :rtype: str
-    """
-    if indata.is_ok():
-        _, no_sign_tx_bytes = indata.result_data
-        return base64.b64encode(base64.b64decode(no_sign_tx_bytes)[:-_GAS_AND_BUDGET_BYTE_OFFSET]).decode()
-    return indata
-
-
-def bcs_from_rpcresult(no_sign_result: SuiRpcResult) -> Union[tuple[str, canoser.Struct], Exception]:
-    """bcs_from_rpcresult converts the transaction bytes from an unsigned transaction result into BCS.
-
-    :param no_sign_result: The result from calling (not signing) transaction
-    :type no_sign_result: SuiRpcResult
-    :raises NotImplementedError: If not implemented by pysui yet
-    :raises AttributeError: Unrecognized TransactionKind
-    :raises AttributeError: No deserialization for BCSSingleTransaction
-    :raises ValueError: If no_sign_result success if False
-    :return: The BCS decoded transaction kind (i.e. Pay, TransferSui, MoveCall, etc.)
-    :rtype: Union[tuple[str, canoser.Struct], Exception]
-    """
-    if no_sign_result.is_ok():
-        _, txbytes = no_sign_result.result_data
-        tx_kind = base64.b64decode(txbytes.value)
-        # The first byte is index of TransactionKind variant (enum)
-        tkind_name, tkind_class = BCSTransactionKind.variant_for_index(tx_kind[_TKIND_INDEX])
-        match tkind_name:
-            case "Single":
-                # The second byte is index into concrete transaction type enum
-                tk_name, tx_type_class = tkind_class.variant_for_index(tx_kind[_SUB_TKIND_INDEX])
-                if tx_type_class:
-                    sheded = tx_kind[:_GAS_AND_BUDGET_BYTE_OFFSET]
-                    sheded = sheded[_SKIP_KIND_AND_SINGLE:]
-                    return tk_name, tx_type_class.deserialize(sheded)
-                raise AttributeError(f"{tkind_name} has no deserialization entry in BCSSingleTransaction")
-            case "Batch":
-                sheded = tx_kind[:_GAS_AND_BUDGET_BYTE_OFFSET]
-                sheded = sheded[_SKIP_KIND_AND_BATCH:]
-                bdser = BCSBatchTransaction.deserialize(sheded)
-                return "Batch", bdser
-            case _:
-                raise AttributeError(f"{tkind_name} is unknown TransactionKind")
-    else:
-        raise ValueError(f"RPC result is invalid {no_sign_result.result_string}")
-
-
 def _bcs_reference_for_oid(client: SyncClient, object_id: ObjectID) -> Union[BCSObjectReference, Exception]:
     """_bcs_reference_for_oid _summary_.
 
-    :param client: _description_
+    :param client: Sui Client
     :type client: SyncClient
-    :param object_id: _description_
+    :param object_id: The object being used for building an ObjectReference
     :type object_id: ObjectID
-    :raises ValueError: _description_
-    :return: _description_
+    :raises ValueError: If sui_getObject fails
+    :return: The BCS ObjectReference
     :rtype: Union[BCSObjectReference, Exception]
     """
     bro_result = client.get_object(object_id)
@@ -404,16 +68,16 @@ def _bcs_reference_for_oid(client: SyncClient, object_id: ObjectID) -> Union[BCS
     raise ValueError(f"{bro_result.result_string} fetching object {object_id}")
 
 
-def _bcs_call_arg_for_oid(client: SyncClient, object_id: ObjectID):
-    """_bcs_call_arg_for_oid generates a call argument for object_id.
+def _bcs_objarg_for_oid(client: SyncClient, object_id: ObjectID) -> Union[ObjectArg, Exception]:
+    """_bcs_objarg_for_oid generates an ObjectArg, used in CallArg, for object_id.
 
-    :param client: _description_
+    :param client: Sui Client
     :type client: SyncClient
-    :param object_id: _description_
+    :param object_id: The object being used for building an ObjectArg
     :type object_id: ObjectID
-    :raises ValueError: If getting object information RPC fails
-    :return: _description_
-    :rtype: _type_
+    :raises ValueError: if sui_getRawObject fails
+    :return: The constructed ObjectArg
+    :rtype: Union[ObjectArg, Exception]
     """
     bro_result = client.execute(GetRawObject(object_id))
     if bro_result.is_ok():
@@ -425,8 +89,140 @@ def _bcs_call_arg_for_oid(client: SyncClient, object_id: ObjectID):
             arg_type = "ImmOrOwnedObject"
         else:
             arg_type = "SharedObject"
-        return CallArg("Object", ObjectArg(arg_type, BCSObjectReference.from_generic_ref(raw_data.reference)))
+        return ObjectArg(arg_type, BCSObjectReference.from_generic_ref(raw_data.reference))
     raise ValueError(f"{bro_result.result_string} fetching object {object_id}")
+
+
+def _bcs_call_arg_for_oid(client: SyncClient, object_id: ObjectID) -> Union[CallArg, Exception]:
+    """_bcs_call_arg_for_oid generates a call argument for object_id.
+
+    :param client: SUI Client
+    :type client: SyncClient
+    :param object_id: The object being used for call arg generation
+    :type object_id: ObjectID
+    :raises ValueError: If getting object information RPC fails
+    :return: A CallArg structure of type "Object"
+    :rtype: CallArg
+    """
+    return CallArg("Object", _bcs_objarg_for_oid(client, object_id))
+
+
+def _bcs_build_pure_array(vector_of: str, in_vector: SuiArray) -> Union[CallArg, Exception]:
+    """_bcs_build_pure_array create a Pure CallArg from array.
+
+    :param vector_of: What contract function is expecting
+    :type vector_of: str
+    :param in_vector: Array of scalars
+    :type in_vector: SuiArray
+    :raises ValueError: If expected type not handled
+    :raises ValueError: If array is not homogeneous types
+    :return: array of u8
+    :rtype: Union[CallArg, Exception]
+    """
+    hmg_type = int
+    call_site = None
+
+    def ints_to_array(byte_count: int, in_el: list[int]) -> list[int]:
+        res: list[int] = []
+        for elem in in_el:
+            byte_res = math.ceil(elem.bit_length() / 8)
+            if byte_res == byte_count:
+                res.extend(list(elem.to_bytes(byte_res, "little")))
+            else:
+                raise ValueError(f"Expected byte count {byte_count} found byte count {byte_res}")
+        return res
+
+    def bools_to_array(in_el: list[bool]) -> list[int]:
+        return [1 if x else 0 for x in in_el]
+
+    def strs_to_array(in_el: list[str]) -> list[int]:
+        res: list[int] = []
+        for elem in in_el:
+            res.extend([ord(x) for x in f"{elem}"])
+        return res
+
+    # Determine expectation of element types
+    if vector_of[0] == "U":
+        hmg_type = int
+        call_site = partial(ints_to_array, int(vector_of[1:]) / 8)
+    elif vector_of[0] == "s":
+        hmg_type = str
+        call_site = strs_to_array
+    elif vector_of[0] == "b":
+        hmg_type = bool
+        call_site = bools_to_array
+    else:
+        raise ValueError(f"Unknown vector type {vector_of}")
+
+    hmg_list = []
+
+    # check for homogenous types and get raw values
+    for elem in in_vector.array:
+        if isinstance(elem.value, hmg_type) is False:
+            raise ValueError(f"{elem.value} of type {type(elem.value)} is not type {vector_of}")
+        hmg_list.append(elem.value)
+    bcs_array = [int.from_bytes(canoser.Uint32.serialize_uint32_as_uleb128(len(hmg_list)), "little")]
+    bcs_array.extend(call_site(hmg_list))
+
+    return CallArg("Pure", bcs_array)
+
+
+def _bcs_call_arguments(
+    client: SyncClient, package_oid: ObjectID, module: SuiString, function: SuiString, arguments: SuiArray
+) -> list[CallArg]:
+    """_bcs_call_arguments generates a list of BCS CallArgs used in Call BCS.
+
+    :param client: SUI Client
+    :type client: SyncClient
+    :param package_oid: ID of published Package
+    :type package_oid: ObjectID
+    :param module: The module name in package
+    :type module: SuiString
+    :param function: The function name in package
+    :type function: SuiString
+    :param arguments: The module arguments list
+    :type arguments: SuiArray
+    :raises ValueError: If length of expected arguments does not match arguments list length
+    :raises ValueError: If expected vector but found other
+    :raises NotImplementedError: If unrecognized function argument meta type
+    :raises ValueError: If sui_getNormalizedMoveFunction fails
+    :return: List of constructed CallArgs
+    :rtype: list[CallArg]
+    """
+    fresult = client.execute(GetFunction(package=package_oid, module_name=module, function_name=function))
+    if fresult.is_ok():
+        arg_list = arguments.arguments
+        func_meta: SuiMoveFunction = fresult.result_data
+        if len(func_meta.parameters) > 0:
+            last_one = func_meta.parameters[-1]
+            meta_parm = last_one.__class__.__name__
+            if meta_parm == "SuiParameterStruct" and last_one.name == "TxContext":
+                func_meta.parameters.pop()
+            elif meta_parm == "SuiParameterReference" and last_one.reference_to.name == "TxContext":
+                func_meta.parameters.pop()
+        if len(func_meta.parameters) != len(arg_list):
+            raise ValueError(f"Function {function} takes {len(func_meta.parameters)} found {len(arg_list)}")
+        call_args: list[CallArg] = []
+        for index, arg in enumerate(arg_list):
+            meta_parm = func_meta.parameters[index]
+            match meta_parm.__class__.__name__:
+                case "SuiParameterReference" | "SuiParameterStruct":
+                    bcs_arg = _bcs_call_arg_for_oid(client, arg)
+                case "SuiMoveScalarArgument":
+                    bcs_arg = CallArg("Pure", [ord(x) for x in f"{arg}"])
+                case "SuiMoveVector":
+                    if isinstance(arg, SuiArray):
+                        if isinstance(meta_parm.vector_of, (SuiParameterReference, SuiParameterStruct)):
+                            bcs_arg = CallArg("ObjVec", [_bcs_objarg_for_oid(client, x) for x in arg.array])
+                        else:
+                            bcs_arg = _bcs_build_pure_array(meta_parm.vector_of, arg)
+                    else:
+                        raise ValueError(f"Expected SuiArray, found {arg.__class__.__name__}")
+                case _:
+                    raise NotImplementedError(f"{func_meta.parameters[index].__class__.__name__} not handled")
+            call_args.append(bcs_arg)
+        return call_args
+    raise ValueError(f"Getting function {function} failed with {fresult.result_string}")
 
 
 def _from_transfer_parms(parms: TransferObjectParams) -> TransferObject:
@@ -467,7 +263,10 @@ def _from_movecall_parms(parms: MoveCallRequestParams) -> MoveCall:
 
 
 def bcs_from_builder(client: SyncClient, builder: _MoveCallTransactionBuilder) -> Union[BCSTransactionKind, Exception]:
-    """."""
+    """bcs_from_builder constructs a Transaction kind from a Builder.
+
+    The result can be serialized and used in sui_devInspectTransaction.
+    """
     bname = builder.__class__.__name__
     tx_kind = "Single"
 
@@ -506,7 +305,9 @@ def bcs_from_builder(client: SyncClient, builder: _MoveCallTransactionBuilder) -
                     TypeTag("Struct", StructTag.from_type_str(x.value)) for x in builder.type_arguments.type_arguments
                 ]
                 # Create list of Call args for any arguments
-                arguments = [_bcs_call_arg_for_oid(client, x.value) for x in builder.arguments.arguments]
+                arguments = _bcs_call_arguments(
+                    client, builder.package_object_id, builder.module, builder.function, builder.arguments
+                )
                 payload = BCSSingleTransaction("Call", BCSMoveCall(package, module, function, type_args, arguments))
             case "BatchTransaction":
                 tx_kind = "Batch"
@@ -530,6 +331,8 @@ def bcs_from_builder(client: SyncClient, builder: _MoveCallTransactionBuilder) -
 def bcs_base64_from_builder(client: SyncClient, builder: _MoveCallTransactionBuilder) -> Union[str, Exception]:
     """bcs_base64_from_builder converts a builder to BCS serialized base64 string.
 
+    Can be then used to submit to sui_devInspectTransaction
+
     :param client: The sui synch client
     :type client: pysui.sui.sui_clients.sync_client.SuiClient
     :param builder: The constructed builder
@@ -538,6 +341,59 @@ def bcs_base64_from_builder(client: SyncClient, builder: _MoveCallTransactionBui
     :rtype: Union[str, Exception]
     """
     return base64.b64encode(bcs_from_builder(client, builder).serialize()).decode()
+
+
+def tkind_from_result(indata: SuiRpcResult) -> Union[str, SuiRpcResult]:
+    """tkind_from_result Return a BCS serialized kind as base64 encoded string.
+
+    Can be then used to submit to sui_devInspectTransaction
+
+    :param indata: The result of calling a complex transaction prior to signing
+    :type indata: SuiRpcResult
+    :return: A base64 encoded serialized BCS TransactionKind
+    :rtype: str
+    """
+    if indata.is_ok():
+        _, no_sign_tx_bytes = indata.result_data
+        return base64.b64encode(base64.b64decode(no_sign_tx_bytes.value)[:_GAS_AND_BUDGET_BYTE_OFFSET]).decode()
+    return indata
+
+
+def bcs_from_rpcresult(no_sign_result: SuiRpcResult) -> Union[tuple[str, canoser.Struct], Exception]:
+    """bcs_from_rpcresult converts the transaction bytes from an unsigned transaction result into BCS.
+
+    :param no_sign_result: The result from calling (not signing) transaction
+    :type no_sign_result: SuiRpcResult
+    :raises NotImplementedError: If not implemented by pysui yet
+    :raises AttributeError: Unrecognized TransactionKind
+    :raises AttributeError: No deserialization for BCSSingleTransaction
+    :raises ValueError: If no_sign_result success if False
+    :return: The BCS decoded transaction kind (i.e. Pay, TransferSui, MoveCall, etc.)
+    :rtype: Union[tuple[str, canoser.Struct], Exception]
+    """
+    if no_sign_result.is_ok():
+        _, txbytes = no_sign_result.result_data
+        tx_kind = base64.b64decode(txbytes.value)
+        # The first byte is index of TransactionKind variant (enum)
+        tkind_name, tkind_class = BCSTransactionKind.variant_for_index(tx_kind[_TKIND_INDEX])
+        match tkind_name:
+            case "Single":
+                # The second byte is index into concrete transaction type enum
+                tk_name, tx_type_class = tkind_class.variant_for_index(tx_kind[_SUB_TKIND_INDEX])
+                if tx_type_class:
+                    sheded = tx_kind[:_GAS_AND_BUDGET_BYTE_OFFSET]
+                    sheded = sheded[_SKIP_KIND_AND_SINGLE:]
+                    return tk_name, tx_type_class.deserialize(sheded)
+                raise AttributeError(f"{tkind_name} has no deserialization entry in BCSSingleTransaction")
+            case "Batch":
+                sheded = tx_kind[:_GAS_AND_BUDGET_BYTE_OFFSET]
+                sheded = sheded[_SKIP_KIND_AND_BATCH:]
+                bdser = BCSBatchTransaction.deserialize(sheded)
+                return "Batch", bdser
+            case _:
+                raise AttributeError(f"{tkind_name} is unknown TransactionKind")
+    else:
+        raise ValueError(f"RPC result is invalid {no_sign_result.result_string}")
 
 
 if __name__ == "__main__":